refactor: BlackPawn, WhitePawn으로 추상화
